Create emptyfolderfinder
This Python application, designed with PyQt5, is called the "Empty Folder Finder." It provides a user interface to find and delete empty directories in your filesystem.

Here's a rundown of its primary features:

Folder Path: This is an input field for the user to specify the directory in which they wish to look for empty folders. This can either be manually typed in or selected via a dialog box which pops up when the user clicks the 'Browse' button.

Maximum Depth: This is a setting that lets the user limit the depth of the search. The user can specify how deep into the directory structure the application should look for empty folders.

Find Empty Folders Button: When this button is pressed, the application starts scanning the specified directory and its subdirectories (up to the specified depth) for empty folders. If no path is provided, the application displays a warning message asking the user to select a folder.

If it finds empty directories, it lists them and asks for user confirmation to delete them. If the user agrees, the application attempts to remove these directories. If any of them can't be deleted, the application will notify the user.

If the search doesn't find any empty directories, the application informs the user that no empty folders were found.

Please note, this application only removes directories that are completely empty, and does not delete directories that contain files or other subdirectories.

This tool can be very useful for cleaning up your file system, especially when working with large repositories where empty folders can inadvertently be left behind.